+++ .gitignore
@@ -1,2 +1,3 @@
 .idea/
 *.iml
+*/target/**
