+++ src/main/java/ru/sbt/mipt/oop/events/EventProduser.java
@@ -1,5 +1,5 @@
 package ru.sbt.mipt.oop.events;
 
 public interface EventProduser {
-    public SensorEvent getNextSensorEvent();
+    SensorEvent getNextSensorEvent();
 }
