+++ src/main/java/ru/sbt/mipt/oop/events/SensorEvent.java
@@ -21,5 +21,4 @@ public abstract class SensorEvent {
 
     public abstract String toString();
 
-    ////public abstract void performEvent(SmartHome smartHome);
 }
