--- src/main/java/ru/sbt/mipt/oop/scenarios/DoorScenario.java
@@ -1,4 +1,0 @@
-package ru.sbt.mipt.oop.scenarios;
-
-public class DoorScenario {
-}
