--- src/main/java/ru/sbt/mipt/oop/scenarios/LightScenarios.java
@@ -1,15 +1,0 @@
-package ru.sbt.mipt.oop.scenarios;
-
-public class LightScenarios implements Scenario {
-/*
-    public static void allLightsOff(SmartHome smartHome) {
-        for (Room homeRoom : smartHome.getRooms()) {
-            for (Light light : homeRoom.getLights()) {
-                light.setOn(false);
-                CommandSender commandSender = new CommandSender();
-                SensorCommand command = new SensorCommand(CommandType.LIGHT_OFF, light.getId());
-                commandSender.sendCommand(command);
-            }
-        }
-    }*/
-}
